refactor mapComponent with utils api calls
